Přejmenování a reorganizace UseCase
