--- Bookong.Application/UseCases/Commands/Interfaces/IAddBookToMaturitaSelectionCommand.cs
@@ -1,10 +1,0 @@
-using System;
-using System.Threading.Tasks;
-
-namespace Bookong.Application.UseCases.Commands.Interfaces
-{
-    public interface IAddBookToMaturitaSelectionCommand
-    {
-        Task ExecuteAsync(Guid bookPublicId);
-    }
-}
