--- Bookong.Application/UseCases/Commands/Interfaces/IRemoveBookFromMaturitaSelectionCommand.cs
@@ -1,10 +1,0 @@
-using System;
-using System.Threading.Tasks;
-
-namespace Bookong.Application.UseCases.Commands.Interfaces
-{
- public interface IRemoveBookFromMaturitaSelectionCommand
- {
- Task ExecuteAsync(Guid bookPublicId);
- }
-}
