+++ Bookong.Application/UseCases/Interfaces/IDeselectMaturitaBookUseCase.cs
@@ -1,0 +1,10 @@
+using Bookong.Application.DTOs;
+using System.Threading.Tasks;
+
+namespace Bookong.Application.UseCases.Interfaces
+{
+    public interface IDeselectMaturitaBookUseCase
+    {
+Task<GenericResponse> Handle(RemoveBookFromPersonalMaturitaListDto dto);
+    }
+}
